added support for config value 'greedy'
By default this collector does greedy matching of interface names, that is, matching eth0 eth1 eth2... when 'eth' is a defined interface.

In our environment this results in collecting a lot of data that we are not interested in as many servers now have a lot of unused interfaces.

Implemented a greedy option so that the interface name matching becomes explicit. The default behavior does not change, but now there is an option for collecting data on only interfaces we care about.