add TotalAccesses metric
ReqPerSec on the apache status page is computed simply by dividing the number
of requests by the uptime of the server.  With a long-running apache instance
this becomes a fairly uninteresting flat average over time.

If we also publish the 'Total Accesses' counter a graphite user can use
derivative(servers.foo.httpd.TotalAccesses) to get a more useful
requests-per-interval rate.