collectors/mongodb: Handle long typed values
With changes to MongoDB version 2.2, `serverStatus` will return
some values as `long` types. For example, here is a snippet of the
result of a `serverStatus` command using MongoDB shell version 2.2.0:

```javascript
	"locks" : {
		"." : {
			"timeLockedMicros" : {
				"R" : NumberLong(200613),
				"W" : NumberLong(113698)
			},
			"timeAcquiringMicros" : {
				"R" : NumberLong(226017),
				"W" : NumberLong(26966)
			}
		},
		"admin" : {
			"timeLockedMicros" : {

			},
			"timeAcquiringMicros" : {

			}
		},
		"local" : {
			"timeLockedMicros" : {
				"r" : NumberLong(3922),
				"w" : NumberLong(14)
			},
			"timeAcquiringMicros" : {
				"r" : NumberLong(941),
				"w" : NumberLong(76)
			}
		}
	}
```

The collector checks that the type is `float` or `int`, so no longer
publishes all values.

This commit checks if the type is `long` and, if so, converts the
value to a `float` before publishing it. It is possible to have
an overflow in this conversion, but that is practically impossible
because the `long`s represent microseconds, whose value should not
exceed that of the range of `float`.