default metric_type to GAUGE
While setting up Librato, I noticed that multiple collectors are not
setting the metric_type parameter (important ones, such as cpu and
network), and thus getting the default of COUNTER. They then getting
recorded wrong and show up as a flat graph. I think COUNTER is the
wrong default, as most metrics collected by Diamond are actually of
GAUGE type. It is also much easier to tell when a counter is accidently
being reported as a guage than when a gauge is being reported as a
counter.