Do not use broken abs mode in threaded interval tasks.
Threading and interval tasks do not mix with abs mode, as the next
job is scheduled immediately after spawning the thread for the current
thread.  Not only does it not make sense (a job running significant
time is scheduled too often), it is actually harmful (a long running
job will cause the calculated delay to be zero, which creates a tight
loop that spawns threads until the system runs out of process slots,
essentially making diamond a temporary fork bomb).