ipvs.py: Add transport specifier to cluster name.
In some cases, for instance when loadbalancing DNS, it may make sense to
have both a UDP and a TCP cluster definitions servicing the same IP and
port on both transport protocols. This was poorly supported before, as
these two clusters' values would be published to the same metric.

This patch prepends all cluster names with their transport protocol and
an underscore, to allow this distinction.