ipvs.py: Also collect current connection counts.
This patch adds collection of the current counts of active and inactive
connections, both per backend and the total counts for the virtual
instance. This depends on an additional run of ipvsadm, as this info
cannot be obtained simultaneously with the --stats argument.

Please note that for setups using Direct Server Return, the active
connection counts will tend to be zero.