Allow -r to run without a full path.
When running diamond in the foreground, it's useful to test single collectors. This functionality already exists, but to extend it:

diamond -f -l -r memorycollector
Will now default to the configured collectors_path and find only collectors that match the name specified.
In this example, if the default collector path is /usr/lib/diamond/collectors and we pass in -r memorycollector, the run_one function will collect all the collectors in /usr/lib/diamond/collectors, remove files that don't match with the name specified in the -r directive, and then execute only those files.