Vagrantfile to build RPM/DEB packages into ./dist/
While trying to update internal version of Diamond, I found the existing
documentation in the Wiki and repo lacking or difficult to map together.
Instead of figuring it out and simply updating the Wiki, I wanted to
provide back an automated process if anyone else needs something
similar. Thus, the Vagrantfile for this project was born. The nodes are
named with a '-build' suffix so as not to conflict with any test VMs
which may be added in the future.