Update diamond
If diamond dies for any reason it leaves the lockfile /var/local/subsys/diamond and the pidfile /var/run/diamond.pid sitting on the disc. This prevents future attempts to 'service diamond start'. This change checks that diamond is still alive before looking at the lockfile. If it is not alive, the lockfile is zapped and it can then be started.