Improvements to the hadoop collector.
This change implements two major changes necessary to use the hadoop
collector in a production environment. The first change divides the time
by 1000 as the time reported in the metrics files are miliseconds since
epoch. Without this change the following error is thrown from graphite:
struct.error: 'L' format requires 0 <= number <= 4294967295

The second change included here is a truncation step to the file read.
Without a truncation step the metrics file will grow forever, and each
run will re-read the whole metrics file, sending all the old, as well as
the new metrics to the graphite server. By truncating the file after a
successful read we can stop not only the disks from filling up, but we
can also ensure that we never overload the graphite server by re-sending
massive numbers of metrics. This change is optional, and defaulted to
false in order to maintain current semantics.