allow to config whether to propagate log of ArchiveHandler to its parent logger
ArchiveHandler is implemented by utilizing logging module, though it is not
considered as a logger. As a handler, it should not log the metrics that
it handles (same as all other handlers do).
This change turns off logging by default and
adds ability to turn that on through its configuration.