added ES aliases check
Added an option that will send metrics of aliased index only... 

One pattern with ES is to alias indices under a static name. re-indexing will be a matter indexing documents in a new index (ex : my_index-2015011302) then change the alias my_index to reference it (and remove the old reference to my_index-2015011301, for example)

This PR will ensure to use the constant alias name as serie name.