follow jaingaurav's suggestion
From my rudimentary reading of the code, I believe the problem is that self.batch is used in the GraphitePickleHandler instead of plain old self.metrics. If you move to using self.metrics I think you will just have to create something like a self._prepare_data function in the GraphiteHandler which either returns the metrics and a string or pickled data depending on the class.

I think if you take that approach you are likely to remove more code than add much.