Allow use of sudo with amavis collector
Typical Debian installations of amavisd create database files (which
amavisd-agent reads) owned amavis:amavis with permissions 0600.

When diamond calls amavisd-agent to read the stats in this setup, it
will fail with a permissions error.

With these changes it's possible to create a sudo configuration which
allows passwordless execution of amavisd-agent by diamond using the
'amavis' user, avoiding this problem.