Filter hadoop collector for whitelist/blacklist python-diamond/Diamond#234
Instead of filtering against the name only (as in collector.publish),
this adds a new collector method to filter the whole path (I believe in
the case of profuse hadoop metrics, this makes sense). The hadoop
collector then calls the new method from it's collect_from method before
publishing. The goal was to enable filtering for this collector without
affecting existing whitelist/blacklist usage.

In the test I have added a metric_blacklist config and then confirmed
that tests broke due to the unpublished metrics. I then removed the
matching metrics from expected.pkl to fix the test. Testing the
whitelist would be a little more challenging.