Send '0' on workers_count metric if process is not found.
This modifies the standard behaviour of not publishing any metrics,
allowing to detect dead processes.

Set process metrics to -1 if the process is not found.

Allow the ProcessResourcesCollector configuration to be split in separate files.

PEP8 fixes

Send -1 on worker count if process is missing and metric requested.

Make missing process reporting as '-1' optional.

Revert "Allow the ProcessResourcesCollector configuration to be split in separate files."

This reverts commit d1f49d5c0dff24ecdac1a601238847c631e80fdc.

Conflicts:
	src/collectors/processresources/processresources.py

Report 0 on worker_count if process not found.

This replaces former widespread -1 reporting as discussed on python-diamond/Diamond#288

Add test for new worker_count functionality.