haproxy.py: Support unix socket collection
In addition to HTTP, HAProxy also exposes statistics through its
configurable admin UNIX domain socket, by use of the command 'show
stat'. This change enables support in the HAProxy collector for using
this for data collection.

To use this, set 'method' to 'unix' in the collector configuration, and
optionally set the socket path with the 'sock' option.

The default behaviour is left unchanged, as the 'method' option defaults
to 'http'.