Force xenfix as opt-in
The bug in Xen which this fix was accommodating was fixed in 2012.
Currently all AWS machines use Xen hypervisors which are already fixed,
however because of the aggressive behavior of the fix, any machines that
are running Diamond on AWS hosts which have not actively configured this
setting are reporting incorrect CPU idle values. This is unfortunate as
this value is often used for provisioning cluster sizes and the like.