Add name/value pair generic file collector.
The intention of this collector is to gather metrics that are found
in files. Candidates would be proc files such as meminfo or vmstat,
but could be used for any file that follows a name/value format.

Given a regex and a file with metrics, it enables you to extract
the name of a metric and its value. It assumes that each line follows
the same format (see vmstat or meminfo). It is very similar to how many
of use awk to extract such values.

This was written in response to the current vmtat collector
limiting the metrics to a very small subset. The idea is that this
can be used for any file, or can be extended to customize behaviour.

The two required config options are 'regex' (which has a defualt value),
and 'file' (with no default value).

Note, the intention of this collector is to be inherited by other
collectors that collect from specific sources, such as various proc
files.