Support HTTPS mod_status URLs and redirects
Previously, HttpdCollector assumed all URLs on their default ports
were requests to port 80. In the case of a `http://localhost/` ->
`https://localhost/` redirect, the latter URL is parsed by `urlparse`
as "Host localhost on the default port" (`parts.netloc` aka `parts[1]`
did not contain a `:`), and a request is made to `http://localhost/`,
which promptly re-redirects to HTTPS again.

Transition to using the attribute names of `urlparse` for readability,
and support both HTTP and HTTPS URLs, by examining the scheme and
constructing the correct httplib class accordingly.