Fixes the missing metrics from kafka.Controller:type=KafkaController
All the metrics under kafka.Controller:type=KafkaController were
missing because Kafka reports the stats as java.lang.Object.
These are critical metrics like OfflinePartitionCount and
ActiveControllerCount.

This patch also handles if the mbeans return empty or None for some
reason. It will log for that particular objectName and move on