Update SignalFX + add filtering
This updates the SignalFX Handler to be able to provide a config
`filter_metrics_regex` that's in the form of comma separated list of
`collector_name:metrics_regex` that can be provided in the config
file.

if this regex isn't set, all metrics are sent, otherwise the metrics
will be filtered based on the list.