Add new 'stagger_collection' option
The stagger functionality was added in 87da1c5. This works, however in
cases where only one collector is configured, it's possible that the
maximum collection time calculated is not long enough and the collection
is terminated with, "Took too long to run!". This is despite the fact
that the configured interval gives more than enough time for the
collector to finish its run.

Therefore, add a new 'stagger_collection' option that stops this
behaviour. When disabled/cleared/set to false, the maximum collection
time will remain static, or in other words it will be set to the
interval configured and will not be increased if the collector doesn't
finish within this time period.

This change could potentially screw up existing deployments that rely on
the existing stagger functionality, so by default 'stagger_collection'
is enabled.