Add replication.master_link_status stat
Redis replicas report the status of their connection to the master, which
can either be "up" or "down". The redisstat collector will now take this
information and publish it as a stat (up: 1, down: 0).

This stat can be used to provide insights into the state of replication,
as well as for alerting on breakage.